+++ packages/uni-app-react/src/dom/mp/events.ts
@@ -203,7 +203,6 @@ export class MPEventTarget {
     options: EventListenerOptions = {}
   ) {
     if (!callback) return
-    // TODO event change render element
 
     const option = normalOptions(options)