+++ packages/uni-app-react/src/hook/index.tsx
@@ -1,0 +1,5 @@
+export * from './event'
+
+// #ifdef MP
+import './mp'
+// #endif