+++ packages/uni-app-react/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @js-css/uni-app-react
 
+## 0.0.5
+
+### Patch Changes
+
+- feat: add core hook
+  fix: preact focus & blur event
+
 ## 0.0.4
 
 ### Patch Changes